Feature/refactor carpetas services (#28)
