+++ src/main/java/com/tecnoinf/gestedu/controllers/AsignaturaController.java
@@ -2,12 +2,10 @@ package com.tecnoinf.gestedu.controllers;
 
 import com.tecnoinf.gestedu.dtos.asignatura.CreateAsignaturaDTO;
 import com.tecnoinf.gestedu.dtos.asignatura.AsignaturaDTO;
-import com.tecnoinf.gestedu.repositories.AsignaturaRepository;
-import com.tecnoinf.gestedu.services.AsignaturaService;
+import com.tecnoinf.gestedu.services.interfaces.AsignaturaService;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
