+++ src/main/java/com/tecnoinf/gestedu/controllers/CarreraController.java
@@ -3,14 +3,13 @@ package com.tecnoinf.gestedu.controllers;
 import com.tecnoinf.gestedu.dtos.carrera.BasicInfoCarreraDTO;
 import com.tecnoinf.gestedu.dtos.carrera.CreateCarreraDTO;
 import com.tecnoinf.gestedu.dtos.asignatura.AsignaturaDTO;
-import com.tecnoinf.gestedu.services.CarreraService;
+import com.tecnoinf.gestedu.services.interfaces.CarreraService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -75,7 +74,7 @@ public class CarreraController {
         return ResponseEntity.ok().body(carreras);
     }
 
-    @Operation(summary = "Obtener las asignaturas de una carrera por su id")
+    @Operation(summary = "Obtener las asignaturas de una carrera por su id (contieen info de plan de estudio si existe el mismo)")
     @GetMapping("/{id}/asignaturas")
     public ResponseEntity<Page<AsignaturaDTO>> getAsignaturasFromCarrera(@PathVariable Long id, Pageable pageable) {
         Page<AsignaturaDTO> asignaturas = carreraService.getAsignaturasFromCarrera(id, pageable);
