+++ src/main/java/com/tecnoinf/gestedu/controllers/DocenteController.java
@@ -1,12 +1,11 @@
 package com.tecnoinf.gestedu.controllers;
 
 import com.tecnoinf.gestedu.dtos.DocenteDTO;
-import com.tecnoinf.gestedu.services.DocenteService;
+import com.tecnoinf.gestedu.services.interfaces.DocenteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
