+++ src/main/java/com/tecnoinf/gestedu/controllers/EstudianteController.java
@@ -1,8 +1,7 @@
 package com.tecnoinf.gestedu.controllers;
 
 import com.tecnoinf.gestedu.dtos.usuario.BasicInfoUsuarioDTO;
-import com.tecnoinf.gestedu.dtos.usuario.UsuarioDTO;
-import com.tecnoinf.gestedu.services.EstudianteService;
+import com.tecnoinf.gestedu.services.interfaces.EstudianteService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
