+++ src/main/java/com/tecnoinf/gestedu/services/interfaces/AsignaturaService.java
@@ -1,4 +1,4 @@
-package com.tecnoinf.gestedu.services;
+package com.tecnoinf.gestedu.services.interfaces;
 
 import com.tecnoinf.gestedu.dtos.asignatura.AsignaturaDTO;
 import com.tecnoinf.gestedu.dtos.asignatura.CreateAsignaturaDTO;
