+++ src/main/java/com/tecnoinf/gestedu/services/interfaces/EstudianteService.java
@@ -1,4 +1,4 @@
-package com.tecnoinf.gestedu.services;
+package com.tecnoinf.gestedu.services.interfaces;
 
 import com.tecnoinf.gestedu.dtos.usuario.BasicInfoUsuarioDTO;
 import org.springframework.data.domain.Page;
